[WIP] abstracting config to env where it makes sense, refactoring elsewhere, adding more tests
